dynamically allocatable array

+++ boundaries.c
@@ -13,10 +13,10 @@ void boundaries(void)
     int iseed2 = 1631;		/* random number seed */
     int i, i_inject;
 
-    float xl = grid_length;	/* shorthand */  
-    float wl = wall_left;
-    float wr = wall_right;
-    float vm, theta, rs;
+    double xl = grid_length;	/* shorthand */  
+    double wl = wall_left;
+    double wr = wall_right;
+    double vm, theta, rs;
 
     /**  loop over all particles to see if any have
          left simulation region: if so, we put them back again
@@ -68,12 +68,12 @@ void boundaries(void)
 
 		/*  find new random velocity in cold Maxwellian */
 
-		rs = (float) random() / random_max;	/* random (0-1) */
+		rs = (double) random() / random_max;	/* random (0-1) */
 
 		i_inject = ne*rs + 1;
 		vm = vte*sqrt(-2.*log((i_inject-0.5)/ne));
 
-		rs = (float) random() / random_max;	/* random (0-1) */
+		rs = (double) random() / random_max;	/* random (0-1) */
 		theta = pi*rs;				/* 1st quadrant */
 		vx[i] = -vm*sin(theta);
 	    }
